Port remaining lifecycle bits to `StableKeyPairDecryptorImpl` and delete `AbstractComponent`
